refactor(engine): a body may import another body, and externs are checked
- internal/body reads the registry into an import table, so a cross-package
  call resolves to the emitted SourcePawn name and needs no hand-written
  extern. attack imports campbomb, and engine.CampBombIsPossible is deleted.
- Generate holds every //sp:body extern's Go signature against the function
  that generates it. Four were wrong: PrintToChatTeam declared TFTeam where
  util.sp says int, GetTravelDistanceToBombTarget declared CNavArea where
  offsets.sp says CTFNavArea, and AimHeadTowards declared its priority and its
  replyWhenAimed as plain ints. All tag drift, none of it visible to spcomp.
- Address_Null and TF2Attrib_GetValue were each declared twice with different
  types. One declaration each, the correctly typed one.
- //sp:library separates a vendored include from SourcePawn the port has not
  reached. 28 of the 34 plugin externs were the former, so the list of work
  remaining was wrong by a factor of five. Five are left.
- engine.Fill puts a panicking stub behind every answer a caller left unset,
  so the nil check leaves 1100 call sites. internal/engine, 18087 to 13265.
- internal/body/slots holds the client array size once. The 41 copies emitted
  41 #defines into one namespace and spcomp warned about 39 every build.
- internal/body/reset.go refuses a per-client array that is neither cleared
  between bots nor //sp:keep-ed. The 45 that already were are listed and
  moved to mvm-z83.91.
- spbody.Cache shares one type check of internal/engine across a run rather
  than one per body. Generate, 9.3s to 0.6s.

The generated SourcePawn moves only where a tag did: a view_as appearing or
disappearing, and the Go_Slots defines going. Both are compile-time only, and
the nav area ones put the output back to what the shipped files wrote.

+++ internal/body/teamchange/teamchange.go
@@ -44,7 +44,7 @@ func HandleTeamPlayerCountChanged(team engine.Team, iWhoChanging int32) {
 	case engine.LineupModeChoose(), engine.LineupModePreferenceChoose():
 		// Allow the classes to be picked again, but do not clear the list.
 		engine.SetBotClassesLocked(false)
-		engine.PrintToChatTeam(team, "%s You can repick your bot team lineup.", engine.PluginPrefix())
+		engine.PrintToChatTeam(int32(team), "%s You can repick your bot team lineup.", engine.PluginPrefix())
 	}
 
 	if !engine.BotsEnabled() {
@@ -54,7 +54,7 @@ func HandleTeamPlayerCountChanged(team engine.Team, iWhoChanging int32) {
 	if iWhoChanging > 0 && engine.ClientOfUserID(engine.BotSummoner()) == iWhoChanging {
 		// The summoner changed teams, so RED may repick its bots.
 		engine.SetAllowBotRedo(true)
-		engine.PrintToChatTeam(team, "%s Use command !redobots to repick your bot team lineup.", engine.PluginPrefix())
+		engine.PrintToChatTeam(int32(team), "%s Use command !redobots to repick your bot team lineup.", engine.PluginPrefix())
 	}
 
 	iWhoToUnready := int32(-1)
